+++ build.borked
@@ -25,14 +25,14 @@ def read_files():
             })
 
 
-def gen_html():
+# def gen_html():
 
-    for p in pages:
-        template = open("./templates/base.html").read()
-        partial = open(p["filename"]).read()
-        #template = template.replace("{{title}}", p["title"])
-        #template = template.replace("{{content}}", partial)
-        open(p["output"], "w+").write(template)
+#     for p in pages:
+#         template = open("./templates/base.html").read()
+#         partial = open(p["filename"]).read()
+#         template = template.replace("{{title}}", p["title"])
+#         template = template.replace("{{content}}", partial)
+#         open(p["output"], "w+").write(template)
 
 
 def gen_blog():
@@ -48,18 +48,18 @@ def gen_blog():
 def read_template():
 
     for p in pages:
-        index_html = open(p["pages"]).read() 
-        template_html = open("templates/base.html").read() 
+        html_file = open(p["filename"]).read() 
+        template_html = open("./templates/base.html").read() 
         template = Template(template_html)
-        template = template.replace("{{title}}", p["title"])
-        open(p["output"], "w+").write(template)
+        template = template.render("{{title}}", p["title"])
+        open(p["output"], "w+").write(html_file)
 
 
 
 def main():
     read_files()
-    #read_template()
-    gen_html()
+    read_template()
+    #gen_html()
     #gen_blog()
     print("Your files have been generated")
