+++ docs/blog.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
-    <title>Blog</title>
+    <title>{{title}}</title>
   		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link href="https://fonts.googleapis.com/css?family=PT+Sans&display=swap" rel="stylesheet"> 
@@ -33,12 +33,7 @@
 				<div class="col-sm-8">
 					<div class="posts">
 						<div>
-													<h1>Kickstart coding, Week 1</h1>
-						<p>10/25/19</p>
-						<p>This week I started my journey at Kickstart coding Bootcamp in Oakland. It is a full stack bootcamp.</p>
-						<p>Some of the technologies we will be learking are: HTML, CSS, JS, Python, Django and much more.</p>
-						<p>The first week we covered html, css and bootstrap.</p><p>Our home work assignment this week is to create our first website/portfolio.</p>
-
+							{{content}}
 						</div>
 					</div>
 				</div> <!-- end col-sm-8 -->
