--- docs/blog4.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-    <title>Blog4</title>
-  		<meta charset="utf-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<link href="https://fonts.googleapis.com/css?family=PT+Sans&display=swap" rel="stylesheet"> 
-		<link rel="stylesheet" href="./css/base.css">
-		<link rel="stylesheet" href="./css/bootstrap.css">
-	</head>
-	<body>
-
-		<div class="container" style="margin-top:30px">
-			<div class="row">
-				<div class="col-sm-4">
-					<div>
-						<img src="./img/rich_houck.jpg" style="border-radius: 50%;"/>
-					</div>
-					<ul class="nav nav-pills flex-column my-custom-nav">
-						<li class="nav-item">
-							<a class="nav-link active" href="index.html">Home</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog1.html">Week1</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog2.html">Week2</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog3.html">Week3</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog4.html">Week4</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="projects.html">Projects</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="mailto:rdhouck@gmail.com">Contact</a>
-						</li>
-					</ul>
-				</div> <!-- end col-sm-4 -->
-        		<div class="col-sm-8">
-					<div class="posts">
-						<div>
-           											<h1>Kickstart coding, Week 4</h1>
-						<p>11/13/19</p>
-            <p>This week we started OOP in python. Our Second Python homework is due</p>
-            <p>This involed our HTML site generator and expanding it to include generating our blog pages</p>
-
-						</div>
-					</div> <!-- end posts -->
-				</div> <!-- end col-sm-8 -->
-			</div> <!-- end row -->
-		</div>
-	</body>
-</html>
