+++ docs/index.bak
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
-    <title>Blog2</title>
+    <title>{{Title}}</title>
   		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link href="https://fonts.googleapis.com/css?family=PT+Sans&display=swap" rel="stylesheet"> 
@@ -9,7 +9,6 @@
 		<link rel="stylesheet" href="./css/bootstrap.css">
 	</head>
 	<body>
-
 		<div class="container" style="margin-top:30px">
 			<div class="row">
 				<div class="col-sm-4">
@@ -21,16 +20,7 @@
 							<a class="nav-link active" href="index.html">Home</a>
 						</li>
 						<li class="nav-item">
-							<a class="nav-link" href="blog1.html">Week1</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog2.html">Week2</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog3.html">Week3</a>
-						</li>
-						<li class="nav-item">
-							<a class="nav-link" href="blog4.html">Week4</a>
+							<a class="nav-link" href="blog1.html">Blog</a>
 						</li>
 						<li class="nav-item">
 							<a class="nav-link" href="projects.html">Projects</a>
@@ -40,19 +30,18 @@
 						</li>
 					</ul>
 				</div> <!-- end col-sm-4 -->
-        		<div class="col-sm-8">
+				<div class="col-sm-8">
 					<div class="posts">
 						<div>
-           												<h1>
-							Kickstart coding, Week 2
-							</h1>
-							<p>10/30/19</p>
-							<p>This week, we stared learning bash and git. I am comfortable with bash as I use it on the job frequently.</p>
-							<p>I use git less frequently and am excited to become more familar with it. I have alread had to use it to revert my project!</p>
-							<p>Our homework this week is involves pushing out code to Github and using a Static site Generator. </p>
+														<h1>About me</h1>
+              <p>Hello My name is Richard. I am a Linux system's Administrator</p>
+              <p> I got my start in the tech industry doing desktop support a while back.<p>
+              <p> The company I worked for was in the  Email marketing space<p><p>I was there 10 years and moved into systems administration about 5 years into the job.<p>
+              <p>I have been doing Linux systems Adminstration for the last 5 years and am enjoying it.<p><p> Most recenltly I have found my self interested in automation and other various technologies.</p>
+              <p>I am currently attending kickstart coding camp in an effort to broaden my skills. I hope to gain the skills there to transition into a devops or developer roll.</p>
 
 						</div>
-					</div> <!-- end posts -->
+					</div>
 				</div> <!-- end col-sm-8 -->
 			</div> <!-- end row -->
 		</div>
