+++ docs/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
-    <title>Index</title>
+    <title>{{title}}</title>
   		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link href="https://fonts.googleapis.com/css?family=PT+Sans&display=swap" rel="stylesheet"> 
@@ -33,13 +33,7 @@
 				<div class="col-sm-8">
 					<div class="posts">
 						<div>
-														<h1>About me</h1>
-              <p>Hello My name is Richard. I am a Linux system's Administrator</p>
-              <p> I got my start in the tech industry doing desktop support a while back.<p>
-              <p> The company I worked for was in the  Email marketing space<p><p>I was there 10 years and moved into systems administration about 5 years into the job.<p>
-              <p>I have been doing Linux systems Adminstration for the last 5 years and am enjoying it.<p><p> Most recenltly I have found my self interested in automation and other various technologies.</p>
-              <p>I am currently attending kickstart coding camp in an effort to broaden my skills. I hope to gain the skills there to transition into a devops or developer roll.</p>
-
+							{{content}}
 						</div>
 					</div>
 				</div> <!-- end col-sm-8 -->
