+++ docs/projects.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 	<head>
-    <title>Projects</title>
+    <title>{{title}}</title>
   		<meta charset="utf-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1">
 		<link href="https://fonts.googleapis.com/css?family=PT+Sans&display=swap" rel="stylesheet"> 
@@ -33,9 +33,7 @@
 				<div class="col-sm-8">
 					<div class="posts">
 						<div>
-														<h1>My first Project</h1>
-							<a href="https://github.com/acanlord/portfolio">link to my portfolio</a>
-
+							{{content}}
 						</div>
 					</div>
 				</div> <!-- end col-sm-8 -->
